src/discord_bot/requirements.txt
discord
boto3
pyyaml
